% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_extension.R
\name{ddbh3_default_conn}
\alias{ddbh3_default_conn}
\title{Get or create default DuckDB connection}
\usage{
ddbh3_default_conn(create = TRUE, upgrade_h3 = FALSE, ...)
}
\arguments{
\item{create}{Logical. If TRUE and no connection exists, create one.
Default is TRUE.}

\item{upgrade_h3}{Logical. If TRUE, will attempt to upgrade the h3 extension}

\item{...}{Additional parameters to pass to \code{\link[duckspatial:ddbs_create_conn]{duckspatial::ddbs_create_conn()}}}
}
\value{
Invisibly, the default \code{duckdb_connection}
}
\description{
Setup the default connection with h3 and spatial extensions installed and loaded.
It will be used internally by the package functions if no other connection is provided
}
\examples{
# Get or create default connection
}
