% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_children_parent.R
\name{ddbh3_get_parent}
\alias{ddbh3_get_parent}
\title{Get the parent H3 cell of H3 indexes}
\usage{
ddbh3_get_parent(
  x,
  resolution,
  h3 = "h3string",
  conn = NULL,
  name = NULL,
  new_column = "h3parent",
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{Input spatial data. Can be:
\itemize{
\item A \code{duckspatial_df} object (lazy spatial data frame via dbplyr)
\item An \code{sf} object
\item A \code{tbl_lazy} from dbplyr
\item A character string naming a table/view in \code{conn}
}
Data is returned from this object.}

\item{resolution}{A number specifying the resolution level of the H3
string (between 0 and 15)}

\item{h3}{The name of a column in \code{x} containing the H3 strings or
H3 unsigned 64-bit integers (\code{UBIGINT})}

\item{conn}{A connection object to a DuckDB database. If \code{NULL}, the function
runs on a temporary DuckDB database.}

\item{name}{A character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table
names. If \code{NULL} (the default), the function returns the result as an
\code{sf} object}

\item{new_column}{Name of the new column to create on the input data. If NULL,
the function will return a vector with the result}

\item{overwrite}{Boolean. whether to overwrite the existing table if it exists. Defaults
to \code{FALSE}. This argument is ignored when \code{name} is \code{NULL}.}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
A \code{tbl_lazy} if \code{x} is not spatial, or a \code{duckspatial_df} if \code{x}
is spatial (e.g. \code{sf} or \code{duckspatial_df}). Alternatively, it creates a table in the connection
if \code{name} is provided, and returns \code{TRUE} invisibly.
}
\description{
Get the parent H3 cell of H3 cell indexes stored as strings or unsigned
64-bit integers (\code{UBIGINT}) at a specified resolution.
}
\examples{
\dontrun{
## Load needed packages
library(duckh3)

## Load example data
points_tbl <- read.csv(
  system.file("extdata/example_pts.csv", package = "duckh3")
)

## Add h3 strings
points_tbl <- ddbh3_lonlat_to_h3(points_tbl, resolution = 10)

## Get resolution-7 parent
points_parent_tbl <- ddbh3_get_parent(points_tbl, 7)

## Check the resolution
ddbh3_get_resolution(
  points_parent_tbl,
  h3 = "h3parent"
)
}
}
