% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_is.R
\name{ddbh3_is}
\alias{ddbh3_is}
\alias{ddbh3_is_pentagon}
\alias{ddbh3_is_valid}
\alias{ddbh3_is_res_class_iii}
\title{Check properties of H3 cell indexes}
\usage{
ddbh3_is_pentagon(
  x,
  h3 = "h3string",
  conn = NULL,
  name = NULL,
  new_column = "ispentagon",
  overwrite = FALSE,
  quiet = FALSE
)

ddbh3_is_valid(
  x,
  h3 = "h3string",
  conn = NULL,
  name = NULL,
  new_column = "isvalid",
  overwrite = FALSE,
  quiet = FALSE
)

ddbh3_is_res_class_iii(
  x,
  h3 = "h3string",
  conn = NULL,
  name = NULL,
  new_column = "isclassiii",
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{Input spatial data. Can be:
\itemize{
\item A \code{duckspatial_df} object (lazy spatial data frame via dbplyr)
\item An \code{sf} object
\item A \code{tbl_lazy} from dbplyr
\item A character string naming a table/view in \code{conn}
}
Data is returned from this object.}

\item{h3}{The name of a column in \code{x} containing the H3 strings or
H3 unsigned 64-bit integers (\code{UBIGINT})}

\item{conn}{A connection object to a DuckDB database. If \code{NULL}, the function
runs on a temporary DuckDB database.}

\item{name}{A character string of length one specifying the name of the table,
or a character string of length two specifying the schema and table
names. If \code{NULL} (the default), the function returns the result as an
\code{sf} object}

\item{new_column}{Name of the new column to create on the input data. If NULL,
the function will return a vector with the result}

\item{overwrite}{Boolean. whether to overwrite the existing table if it exists. Defaults
to \code{FALSE}. This argument is ignored when \code{name} is \code{NULL}.}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
A \code{tbl_lazy} if \code{x} is not spatial, or a \code{duckspatial_df} if \code{x}
is spatial (e.g. \code{sf} or \code{duckspatial_df}). Alternatively, it creates a table in the connection
if \code{name} is provided, and returns \code{TRUE} invisibly.
}
\description{
Check properties of H3 cell indexes stored as strings or unsigned 64-bit
integers
}
\details{
The three functions check different properties of H3 cell indexes, all
returning a logical column:
\itemize{
\item \code{ddbh3_is_pentagon()} returns \code{TRUE} if the H3 cell is one of the 12
pentagonal cells that exist at each H3 resolution
\item \code{ddbh3_is_valid()} returns \code{TRUE} if the H3 cell index is a valid H3 cell
\item \code{ddbh3_is_res_class_iii()} returns \code{TRUE} if the H3 cell belongs to a
Class III resolution (odd resolutions: 1, 3, 5, 7, 9, 11, 13, 15)
}
}
\examples{
\dontrun{
## TODO
}
}
